requirements.txt
gTTS==2.2.3
pyTelegramBotAPI==4.2.1
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
zipp==3.6.0
python-dotenv==1.0.1
